Add regression test for changing name of simple product
