+++ src/Sylius/Behat/Context/Ui/Admin/ManagingProductVariantsContext.php
@@ -22,7 +22,6 @@ use Sylius\Behat\Service\Resolver\CurrentPageResolverInterface;
 use Sylius\Behat\Service\SharedStorageInterface;
 use Sylius\Component\Core\Model\ProductInterface;
 use Sylius\Component\Core\Model\ProductVariantInterface;
-use Sylius\Component\Product\Resolver\ProductVariantResolverInterface;
 use Webmozart\Assert\Assert;
 
 /**
@@ -35,11 +34,6 @@ final class ManagingProductVariantsContext implements Context
      * @var SharedStorageInterface
      */
     private $sharedStorage;
-
-    /**
-     * @var ProductVariantResolverInterface
-     */
-    private $defaultProductVariantResolver;
 
     /**
      * @var CreatePageInterface
@@ -73,7 +67,6 @@ final class ManagingProductVariantsContext implements Context
 
     /**
      * @param SharedStorageInterface $sharedStorage
-     * @param ProductVariantResolverInterface $defaultProductVariantResolver
      * @param CreatePageInterface $createPage
      * @param IndexPageInterface $indexPage
      * @param UpdatePageInterface $updatePage
@@ -83,7 +76,6 @@ final class ManagingProductVariantsContext implements Context
      */
     public function __construct(
         SharedStorageInterface $sharedStorage,
-        ProductVariantResolverInterface $defaultProductVariantResolver,
         CreatePageInterface $createPage,
         IndexPageInterface $indexPage,
         UpdatePageInterface $updatePage,
@@ -92,7 +84,6 @@ final class ManagingProductVariantsContext implements Context
         NotificationCheckerInterface $notificationChecker
     ) {
         $this->sharedStorage = $sharedStorage;
-        $this->defaultProductVariantResolver = $defaultProductVariantResolver;
         $this->createPage = $createPage;
         $this->indexPage = $indexPage;
         $this->updatePage = $updatePage;
@@ -209,14 +200,6 @@ final class ManagingProductVariantsContext implements Context
     }
 
     /**
-     * @When I start sorting variants by :field
-     */
-    public function iSortProductsBy($field)
-    {
-        $this->indexPage->sortBy($field);
-    }
-
-    /**
      * @When I set the position of :name to :position
      */
     public function iSetThePositionOfTo($name, $position)
@@ -230,42 +213,6 @@ final class ManagingProductVariantsContext implements Context
     public function iSaveMyNewConfiguration()
     {
         $this->indexPage->savePositions();
-    }
-
-    /**
-     * @Then the :productVariantCode variant of the :product product should appear in the store
-     */
-    public function theProductVariantShouldAppearInTheShop($productVariantCode, ProductInterface $product)
-    {
-        $this->iWantToViewAllVariantsOfThisProduct($product);
-
-        Assert::true($this->indexPage->isSingleResourceOnPage(['code' => $productVariantCode]));
-    }
-
-    /**
-     * @Then the :productVariantCode variant of the :product product should not appear in the store
-     */
-    public function theProductVariantShouldNotAppearInTheShop($productVariantCode, ProductInterface $product)
-    {
-        $this->iWantToViewAllVariantsOfThisProduct($product);
-
-        Assert::false($this->indexPage->isSingleResourceOnPage(['code' => $productVariantCode]));
-    }
-
-    /**
-     * @Then the :product product should have no variants
-     */
-    public function theProductShouldHaveNoVariants(ProductInterface $product)
-    {
-        $this->assertNumberOfVariantsOnProductPage($product, 0);
-    }
-
-    /**
-     * @Then the :product product should have only one variant
-     */
-    public function theProductShouldHaveOnlyOneVariant(ProductInterface $product)
-    {
-        $this->assertNumberOfVariantsOnProductPage($product, 1);
     }
 
     /**
@@ -302,43 +249,14 @@ final class ManagingProductVariantsContext implements Context
     }
 
     /**
-     * @When /^I (?:|want to )view all variants of (this product)$/
-     * @When /^I view(?:| all) variants of the (product "[^"]+")$/
-     */
-    public function iWantToViewAllVariantsOfThisProduct(ProductInterface $product)
-    {
-        $this->indexPage->open(['productId' => $product->getId()]);
-    }
-
-    /**
-     * @Then I should see :numberOfProductVariants variants in the list
-     * @Then I should see :numberOfProductVariants variant in the list
-     * @Then I should not see any variants in the list
-     */
-    public function iShouldSeeProductVariantsInTheList($numberOfProductVariants = 0)
-    {
-        Assert::same($this->indexPage->countItems(), (int) $numberOfProductVariants);
-    }
-
-    /**
      * @When /^I delete the ("[^"]+" variant of product "[^"]+")$/
      * @When /^I try to delete the ("[^"]+" variant of product "[^"]+")$/
      */
     public function iDeleteTheVariantOfProduct(ProductVariantInterface $productVariant)
     {
-        $this->iWantToViewAllVariantsOfThisProduct($productVariant->getProduct());
+        $this->indexPage->open(['productId' => $productVariant->getProduct()->getId()]);
 
         $this->indexPage->deleteResourceOnPage(['code' => $productVariant->getCode()]);
-    }
-
-    /**
-     * @Then /^(this variant) should not exist in the product catalog$/
-     */
-    public function productVariantShouldNotExist(ProductVariantInterface $productVariant)
-    {
-        $this->iWantToViewAllVariantsOfThisProduct($productVariant->getProduct());
-
-        Assert::false($this->indexPage->isSingleResourceOnPage(['name' => $productVariant->getName()]));
     }
 
     /**
@@ -350,14 +268,6 @@ final class ManagingProductVariantsContext implements Context
             'Cannot delete, the product variant is in use.',
             NotificationType::failure()
         );
-    }
-
-    /**
-     * @Then /^(this variant) should still exist in the product catalog$/
-     */
-    public function productShouldExistInTheProductCatalog(ProductVariantInterface $productVariant)
-    {
-        $this->theProductVariantShouldAppearInTheShop($productVariant->getCode(), $productVariant->getProduct());
     }
 
     /**
@@ -495,30 +405,6 @@ final class ManagingProductVariantsContext implements Context
     }
 
     /**
-     * @Then /^the variant "([^"]+)" should have (\d+) items on hand$/
-     */
-    public function thisVariantShouldHaveItemsOnHand($productVariantName, $quantity)
-    {
-        Assert::true($this->indexPage->isSingleResourceWithSpecificElementOnPage(
-            ['name' => $productVariantName],
-            sprintf('td > div.ui.label:contains("%s")', $quantity)
-        ));
-    }
-
-    /**
-     * @Then /^the "([^"]+)" variant of ("[^"]+" product) should have (\d+) items on hand$/
-     */
-    public function theVariantOfProductShouldHaveItemsOnHand($productVariantName, ProductInterface $product, $quantity)
-    {
-        $this->indexPage->open(['productId' => $product->getId()]);
-
-        Assert::true($this->indexPage->isSingleResourceWithSpecificElementOnPage(
-            ['name' => $productVariantName],
-            sprintf('td > div.ui.label:contains("%s")', $quantity)
-        ));
-    }
-
-    /**
      * @Then /^inventory of (this variant) should not be tracked$/
      */
     public function thisProductVariantShouldNotBeTracked(ProductVariantInterface $productVariant)
@@ -536,97 +422,6 @@ final class ManagingProductVariantsContext implements Context
         $this->iWantToModifyAProduct($productVariant);
 
         Assert::true($this->updatePage->isTracked());
-    }
-
-    /**
-     * @Then /^I should see that the ("([^"]+)" variant) is not tracked$/
-     */
-    public function iShouldSeeThatIsNotTracked(ProductVariantInterface $productVariant)
-    {
-        Assert::true($this->indexPage->isSingleResourceOnPage([
-            'name' => $productVariant->getName(),
-            'inventory' => 'Not tracked',
-        ]));
-    }
-
-    /**
-     * @Then /^I should see that the ("[^"]+" variant) has zero on hand quantity$/
-     */
-    public function iShouldSeeThatTheVariantHasZeroOnHandQuantity(ProductVariantInterface $productVariant)
-    {
-        Assert::true($this->indexPage->isSingleResourceOnPage([
-            'name' => $productVariant->getName(),
-            'inventory' => '0 Available on hand',
-        ]));
-    }
-
-    /**
-     * @Then /^(\d+) units of (this product) should be on hold$/
-     */
-    public function unitsOfThisProductShouldBeOnHold($quantity, ProductInterface $product)
-    {
-        /** @var ProductVariantInterface $variant */
-        $variant = $this->defaultProductVariantResolver->getVariant($product);
-
-        $this->assertOnHoldQuantityOfVariant($quantity, $variant);
-    }
-
-    /**
-     * @Then /^(\d+) units of (this product) should be on hand$/
-     */
-    public function unitsOfThisProductShouldBeOnHand($quantity, ProductInterface $product)
-    {
-        /** @var ProductVariantInterface $variant */
-        $variant = $this->defaultProductVariantResolver->getVariant($product);
-
-        Assert::same($this->indexPage->getOnHandQuantityFor($variant), (int) $quantity);
-    }
-
-    /**
-     * @Then /^there should be no units of (this product) on hold$/
-     */
-    public function thereShouldBeNoUnitsOfThisProductOnHold(ProductInterface $product)
-    {
-        /** @var ProductVariantInterface $variant */
-        $variant = $this->defaultProductVariantResolver->getVariant($product);
-
-        $this->assertOnHoldQuantityOfVariant(0, $variant);
-    }
-
-    /**
-     * @Then the :variant variant should have :amount items on hold
-     */
-    public function thisVariantShouldHaveItemsOnHold(ProductVariantInterface $variant, $amount)
-    {
-        $this->assertOnHoldQuantityOfVariant((int) $amount, $variant);
-    }
-
-    /**
-     * @Then the :variant variant of :product product should have :amount items on hold
-     */
-    public function theVariantOfProductShouldHaveItemsOnHold(ProductVariantInterface $variant, ProductInterface $product, $amount)
-    {
-        $this->indexPage->open(['productId' => $product->getId()]);
-
-        $this->assertOnHoldQuantityOfVariant((int) $amount, $variant);
-    }
-
-    /**
-     * @Then the first variant in the list should have :field :value
-     */
-    public function theFirstVariantInTheListShouldHave($field, $value)
-    {
-        Assert::same($this->indexPage->getColumnFields($field)[0], $value);
-    }
-
-    /**
-     * @Then the last variant in the list should have :field :value
-     */
-    public function theLastVariantInTheListShouldHave($field, $value)
-    {
-        $values = $this->indexPage->getColumnFields($field);
-
-        Assert::same(end($values), $value);
     }
 
     /**
@@ -712,16 +507,6 @@ final class ManagingProductVariantsContext implements Context
     }
 
     /**
-     * @Then /^(this variant) should have a (\d+) item currently in stock$/
-     */
-    public function thisVariantShouldHaveAItemCurrentlyInStock(ProductVariantInterface $productVariant, $amountInStock)
-    {
-        $this->indexPage->open(['productId' => $productVariant->getProduct()->getId()]);
-
-        Assert::same($this->indexPage->getOnHandQuantityFor($productVariant), (int) $amountInStock);
-    }
-
-    /**
      * @Then I should be notified that on hand quantity must be greater than the number of on hold units
      */
     public function iShouldBeNotifiedThatOnHandQuantityMustBeGreaterThanTheNumberOfOnHoldUnits()
@@ -742,38 +527,5 @@ final class ManagingProductVariantsContext implements Context
         $currentPage = $this->currentPageResolver->getCurrentPageWithForm([$this->createPage, $this->updatePage]);
 
         Assert::same($currentPage->getValidationMessage($element), $message);
-    }
-
-    /**
-     * @param int $expectedAmount
-     * @param ProductVariantInterface $variant
-     *
-     * @throws \InvalidArgumentException
-     */
-    private function assertOnHoldQuantityOfVariant($expectedAmount, $variant)
-    {
-        $actualAmount = $this->indexPage->getOnHoldQuantityFor($variant);
-
-        Assert::same(
-            $actualAmount,
-            (int) $expectedAmount,
-            sprintf(
-                'Unexpected on hold quantity for "%s" variant. It should be "%s" but is "%s"',
-                $variant->getName(),
-                $expectedAmount,
-                $actualAmount
-            )
-        );
-    }
-
-    /**
-     * @param ProductInterface $product
-     * @param int $amount
-     */
-    private function assertNumberOfVariantsOnProductPage(ProductInterface $product, $amount)
-    {
-        $this->iWantToViewAllVariantsOfThisProduct($product);
-
-        Assert::same((int) $this->indexPage->countItems(), $amount, 'Product has %d variants, but should have %d');
     }
 }
