+++ src/Sylius/Behat/Resources/config/suites/ui/inventory/checkout_inventory.yml
@@ -25,6 +25,7 @@ default:
                 - sylius.behat.context.transform.shared_storage
                 - sylius.behat.context.transform.shipping_method
 
+                - sylius.behat.context.ui.admin.browsing_product_variants
                 - sylius.behat.context.ui.admin.managing_product_variants
                 - sylius.behat.context.ui.shop.cart
                 - sylius.behat.context.ui.shop.checkout
