+++ src/Sylius/Behat/Resources/config/suites/ui/product/managing_products.yml
@@ -40,6 +40,7 @@ default:
                 - sylius.behat.context.setup.taxonomy
                 - sylius.behat.context.setup.zone
 
+                - sylius.behat.context.ui.admin.browsing_product_variants
                 - sylius.behat.context.ui.admin.managing_products
                 - sylius.behat.context.ui.admin.notification
                 - sylius.behat.context.ui.shop.locale
